+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/db_attributes.py
@@ -12,335 +12,383 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 from enum import Enum
+from typing import Final
 
-DB_CASSANDRA_CONSISTENCY_LEVEL = "db.cassandra.consistency_level"
+DB_CASSANDRA_CONSISTENCY_LEVEL: Final = "db.cassandra.consistency_level"
 """
 The consistency level of the query. Based on consistency values from [CQL](https://docs.datastax.com/en/cassandra-oss/3.0/cassandra/dml/dmlConfigConsistency.html).
 """
 
-DB_CASSANDRA_COORDINATOR_DC = "db.cassandra.coordinator.dc"
+DB_CASSANDRA_COORDINATOR_DC: Final = "db.cassandra.coordinator.dc"
 """
 The data center of the coordinating node for a query.
 """
 
-DB_CASSANDRA_COORDINATOR_ID = "db.cassandra.coordinator.id"
+DB_CASSANDRA_COORDINATOR_ID: Final = "db.cassandra.coordinator.id"
 """
 The ID of the coordinating node for a query.
 """
 
-DB_CASSANDRA_IDEMPOTENCE = "db.cassandra.idempotence"
+DB_CASSANDRA_IDEMPOTENCE: Final = "db.cassandra.idempotence"
 """
 Whether or not the query is idempotent.
 """
 
-DB_CASSANDRA_PAGE_SIZE = "db.cassandra.page_size"
+DB_CASSANDRA_PAGE_SIZE: Final = "db.cassandra.page_size"
 """
 The fetch size used for paging, i.e. how many rows will be returned at once.
 """
 
-DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT = (
+DB_CASSANDRA_SPECULATIVE_EXECUTION_COUNT: Final = (
     "db.cassandra.speculative_execution_count"
 )
 """
 The number of times a query was speculatively executed. Not set or `0` if the query was not executed speculatively.
 """
 
-DB_CASSANDRA_TABLE = "db.cassandra.table"
+DB_CASSANDRA_TABLE: Final = "db.cassandra.table"
 """
-The name of the primary Cassandra table that the operation is acting upon, including the keyspace name (if applicable).
-Note: This mirrors the db.sql.table attribute but references cassandra rather than sql. It is not recommended to attempt any client-side parsing of `db.statement` just to get this property, but it should be set if it is provided by the library being instrumented. If the operation is acting upon an anonymous table, or more than one table, this value MUST NOT be set.
+Deprecated: Replaced by `db.collection.name`.
 """
 
-DB_CONNECTION_STRING = "db.connection_string"
+DB_CLIENT_CONNECTIONS_POOL_NAME: Final = "db.client.connections.pool.name"
+"""
+The name of the connection pool; unique within the instrumented application. In case the connection pool implementation doesn't provide a name, instrumentation should use a combination of `server.address` and `server.port` attributes formatted as `server.address:server.port`.
+"""
+
+DB_CLIENT_CONNECTIONS_STATE: Final = "db.client.connections.state"
+"""
+The state of a connection in the pool.
+"""
+
+DB_COLLECTION_NAME: Final = "db.collection.name"
+"""
+The name of a collection (table, container) within the database.
+Note: If the collection name is parsed from the query, it SHOULD match the value provided in the query and may be qualified with the schema and database name.
+    It is RECOMMENDED to capture the value as provided by the application without attempting to do any case normalization.
+"""
+
+DB_CONNECTION_STRING: Final = "db.connection_string"
 """
 Deprecated: "Replaced by `server.address` and `server.port`.".
 """
 
-DB_COSMOSDB_CLIENT_ID = "db.cosmosdb.client_id"
+DB_COSMOSDB_CLIENT_ID: Final = "db.cosmosdb.client_id"
 """
 Unique Cosmos client instance id.
 """
 
-DB_COSMOSDB_CONNECTION_MODE = "db.cosmosdb.connection_mode"
+DB_COSMOSDB_CONNECTION_MODE: Final = "db.cosmosdb.connection_mode"
 """
 Cosmos client connection mode.
 """
 
-DB_COSMOSDB_CONTAINER = "db.cosmosdb.container"
+DB_COSMOSDB_CONTAINER: Final = "db.cosmosdb.container"
 """
-Cosmos DB container name.
+Deprecated: Replaced by `db.collection.name`.
 """
 
-DB_COSMOSDB_OPERATION_TYPE = "db.cosmosdb.operation_type"
+DB_COSMOSDB_OPERATION_TYPE: Final = "db.cosmosdb.operation_type"
 """
 CosmosDB Operation Type.
 """
 
-DB_COSMOSDB_REQUEST_CHARGE = "db.cosmosdb.request_charge"
+DB_COSMOSDB_REQUEST_CHARGE: Final = "db.cosmosdb.request_charge"
 """
 RU consumed for that operation.
 """
 
-DB_COSMOSDB_REQUEST_CONTENT_LENGTH = "db.cosmosdb.request_content_length"
+DB_COSMOSDB_REQUEST_CONTENT_LENGTH: Final = (
+    "db.cosmosdb.request_content_length"
+)
 """
 Request payload size in bytes.
 """
 
-DB_COSMOSDB_STATUS_CODE = "db.cosmosdb.status_code"
+DB_COSMOSDB_STATUS_CODE: Final = "db.cosmosdb.status_code"
 """
 Cosmos DB status code.
 """
 
-DB_COSMOSDB_SUB_STATUS_CODE = "db.cosmosdb.sub_status_code"
+DB_COSMOSDB_SUB_STATUS_CODE: Final = "db.cosmosdb.sub_status_code"
 """
 Cosmos DB sub status code.
 """
 
-DB_ELASTICSEARCH_CLUSTER_NAME = "db.elasticsearch.cluster.name"
+DB_ELASTICSEARCH_CLUSTER_NAME: Final = "db.elasticsearch.cluster.name"
 """
 Represents the identifier of an Elasticsearch cluster.
 """
 
-DB_ELASTICSEARCH_NODE_NAME = "db.elasticsearch.node.name"
+DB_ELASTICSEARCH_NODE_NAME: Final = "db.elasticsearch.node.name"
 """
-Deprecated: Replaced by `db.instance.id`.
+Represents the human-readable identifier of the node/instance to which a request was routed.
 """
 
-DB_ELASTICSEARCH_PATH_PARTS_TEMPLATE = "db.elasticsearch.path_parts"
+DB_ELASTICSEARCH_PATH_PARTS_TEMPLATE: Final = "db.elasticsearch.path_parts"
 """
 A dynamic value in the url path.
 Note: Many Elasticsearch url paths allow dynamic values. These SHOULD be recorded in span attributes in the format `db.elasticsearch.path_parts.<key>`, where `<key>` is the url path part name. The implementation SHOULD reference the [elasticsearch schema](https://raw.githubusercontent.com/elastic/elasticsearch-specification/main/output/schema/schema.json) in order to map the path part values to their names.
 """
 
-DB_INSTANCE_ID = "db.instance.id"
+DB_INSTANCE_ID: Final = "db.instance.id"
 """
-An identifier (address, unique name, or any other identifier) of the database instance that is executing queries or mutations on the current connection. This is useful in cases where the database is running in a clustered environment and the instrumentation is able to record the node executing the query. The client may obtain this value in databases like MySQL using queries like `select @@hostname`.
+Deprecated: Deprecated, no general replacement at this time. For Elasticsearch, use `db.elasticsearch.node.name` instead.
 """
 
-DB_JDBC_DRIVER_CLASSNAME = "db.jdbc.driver_classname"
+DB_JDBC_DRIVER_CLASSNAME: Final = "db.jdbc.driver_classname"
 """
 Deprecated: Removed as not used.
 """
 
-DB_MONGODB_COLLECTION = "db.mongodb.collection"
+DB_MONGODB_COLLECTION: Final = "db.mongodb.collection"
 """
-The MongoDB collection being accessed within the database stated in `db.name`.
-"""
-
-DB_MSSQL_INSTANCE_NAME = "db.mssql.instance_name"
-"""
-The Microsoft SQL Server [instance name](https://docs.microsoft.com/sql/connect/jdbc/building-the-connection-url?view=sql-server-ver15) connecting to. This name is used to determine the port of a named instance.
-Note: If setting a `db.mssql.instance_name`, `server.port` is no longer required (but still recommended if non-standard).
+Deprecated: Replaced by `db.collection.name`.
 """
 
-DB_NAME = "db.name"
+DB_MSSQL_INSTANCE_NAME: Final = "db.mssql.instance_name"
 """
-This attribute is used to report the name of the database being accessed. For commands that switch the database, this should be set to the target database (even if the command fails).
-Note: In some SQL databases, the database name to be used is called "schema name". In case there are multiple layers that could be considered for database name (e.g. Oracle instance name and schema name), the database name to be used is the more specific layer (e.g. Oracle schema name).
-"""
-
-DB_OPERATION = "db.operation"
-"""
-The name of the operation being executed, e.g. the [MongoDB command name](https://docs.mongodb.com/manual/reference/command/#database-operations) such as `findAndModify`, or the SQL keyword.
-Note: When setting this to an SQL keyword, it is not recommended to attempt any client-side parsing of `db.statement` just to get this property, but it should be set if the operation name is provided by the library being instrumented. If the SQL statement has an ambiguous operation, or performs more than one operation, this value may be omitted.
+Deprecated: Deprecated, no replacement at this time.
 """
 
-DB_REDIS_DATABASE_INDEX = "db.redis.database_index"
+DB_NAME: Final = "db.name"
 """
-The index of the database being accessed as used in the [`SELECT` command](https://redis.io/commands/select), provided as an integer. To be used instead of the generic `db.name` attribute.
-"""
-
-DB_SQL_TABLE = "db.sql.table"
-"""
-The name of the primary table that the operation is acting upon, including the database name (if applicable).
-Note: It is not recommended to attempt any client-side parsing of `db.statement` just to get this property, but it should be set if it is provided by the library being instrumented. If the operation is acting upon an anonymous table, or more than one table, this value MUST NOT be set.
+Deprecated: Replaced by `db.namespace`.
 """
 
-DB_STATEMENT = "db.statement"
+DB_NAMESPACE: Final = "db.namespace"
 """
-The database statement being executed.
-"""
-
-DB_SYSTEM = "db.system"
-"""
-An identifier for the database management system (DBMS) product being used. See below for a list of well-known identifiers.
+The name of the database, fully qualified within the server address and port.
+Note: If a database system has multiple namespace components, they SHOULD be concatenated (potentially using database system specific conventions) from most general to most specific namespace component, and more specific namespaces SHOULD NOT be captured without the more general namespaces, to ensure that "startswith" queries for the more general namespaces will be valid.
+    Semantic conventions for individual database systems SHOULD document what `db.namespace` means in the context of that system.
+    It is RECOMMENDED to capture the value as provided by the application without attempting to do any case normalization.
 """
 
-DB_USER = "db.user"
+DB_OPERATION: Final = "db.operation"
 """
-Username for accessing the database.
+Deprecated: Replaced by `db.operation.name`.
+"""
+
+DB_OPERATION_NAME: Final = "db.operation.name"
+"""
+The name of the operation or command being executed.
+Note: It is RECOMMENDED to capture the value as provided by the application without attempting to do any case normalization.
+"""
+
+DB_QUERY_PARAMETER_TEMPLATE: Final = "db.query.parameter"
+"""
+The query parameters used in `db.query.text`, with `<key>` being the parameter name, and the attribute value being the parameter value.
+Note: Query parameters should only be captured when `db.query.text` is parameterized with placeholders.
+    If a parameter has no name and instead is referenced only by index, then `<key>` SHOULD be the 0-based index.
+"""
+
+DB_QUERY_TEXT: Final = "db.query.text"
+"""
+The database query being executed.
+"""
+
+DB_REDIS_DATABASE_INDEX: Final = "db.redis.database_index"
+"""
+Deprecated: Replaced by `db.namespace`.
+"""
+
+DB_SQL_TABLE: Final = "db.sql.table"
+"""
+Deprecated: Replaced by `db.collection.name`.
+"""
+
+DB_STATEMENT: Final = "db.statement"
+"""
+Deprecated: Replaced by `db.query.text`.
+"""
+
+DB_SYSTEM: Final = "db.system"
+"""
+The database management system (DBMS) product as identified by the client instrumentation.
+Note: The actual DBMS may differ from the one identified by the client. For example, when using PostgreSQL client libraries to connect to a CockroachDB, the `db.system` is set to `postgresql` based on the instrumentation's best knowledge.
+"""
+
+DB_USER: Final = "db.user"
+"""
+Deprecated: No replacement at this time.
 """
 
 
 class DbCassandraConsistencyLevelValues(Enum):
-    ALL = "all"
+    ALL: Final = "all"
     """all."""
-    EACH_QUORUM = "each_quorum"
+    EACH_QUORUM: Final = "each_quorum"
     """each_quorum."""
-    QUORUM = "quorum"
+    QUORUM: Final = "quorum"
     """quorum."""
-    LOCAL_QUORUM = "local_quorum"
+    LOCAL_QUORUM: Final = "local_quorum"
     """local_quorum."""
-    ONE = "one"
+    ONE: Final = "one"
     """one."""
-    TWO = "two"
+    TWO: Final = "two"
     """two."""
-    THREE = "three"
+    THREE: Final = "three"
     """three."""
-    LOCAL_ONE = "local_one"
+    LOCAL_ONE: Final = "local_one"
     """local_one."""
-    ANY = "any"
+    ANY: Final = "any"
     """any."""
-    SERIAL = "serial"
+    SERIAL: Final = "serial"
     """serial."""
-    LOCAL_SERIAL = "local_serial"
+    LOCAL_SERIAL: Final = "local_serial"
     """local_serial."""
 
 
+class DbClientConnectionsStateValues(Enum):
+    IDLE: Final = "idle"
+    """idle."""
+    USED: Final = "used"
+    """used."""
+
+
 class DbCosmosdbConnectionModeValues(Enum):
-    GATEWAY = "gateway"
+    GATEWAY: Final = "gateway"
     """Gateway (HTTP) connections mode."""
-    DIRECT = "direct"
+    DIRECT: Final = "direct"
     """Direct connection."""
 
 
 class DbCosmosdbOperationTypeValues(Enum):
-    INVALID = "Invalid"
+    INVALID: Final = "Invalid"
     """invalid."""
-    CREATE = "Create"
+    CREATE: Final = "Create"
     """create."""
-    PATCH = "Patch"
+    PATCH: Final = "Patch"
     """patch."""
-    READ = "Read"
+    READ: Final = "Read"
     """read."""
-    READ_FEED = "ReadFeed"
+    READ_FEED: Final = "ReadFeed"
     """read_feed."""
-    DELETE = "Delete"
+    DELETE: Final = "Delete"
     """delete."""
-    REPLACE = "Replace"
+    REPLACE: Final = "Replace"
     """replace."""
-    EXECUTE = "Execute"
+    EXECUTE: Final = "Execute"
     """execute."""
-    QUERY = "Query"
+    QUERY: Final = "Query"
     """query."""
-    HEAD = "Head"
+    HEAD: Final = "Head"
     """head."""
-    HEAD_FEED = "HeadFeed"
+    HEAD_FEED: Final = "HeadFeed"
     """head_feed."""
-    UPSERT = "Upsert"
+    UPSERT: Final = "Upsert"
     """upsert."""
-    BATCH = "Batch"
+    BATCH: Final = "Batch"
     """batch."""
-    QUERY_PLAN = "QueryPlan"
+    QUERY_PLAN: Final = "QueryPlan"
     """query_plan."""
-    EXECUTE_JAVASCRIPT = "ExecuteJavaScript"
+    EXECUTE_JAVASCRIPT: Final = "ExecuteJavaScript"
     """execute_javascript."""
 
 
 class DbSystemValues(Enum):
-    OTHER_SQL = "other_sql"
+    OTHER_SQL: Final = "other_sql"
     """Some other SQL database. Fallback only. See notes."""
-    MSSQL = "mssql"
+    MSSQL: Final = "mssql"
     """Microsoft SQL Server."""
-    MSSQLCOMPACT = "mssqlcompact"
+    MSSQLCOMPACT: Final = "mssqlcompact"
     """Microsoft SQL Server Compact."""
-    MYSQL = "mysql"
+    MYSQL: Final = "mysql"
     """MySQL."""
-    ORACLE = "oracle"
+    ORACLE: Final = "oracle"
     """Oracle Database."""
-    DB2 = "db2"
+    DB2: Final = "db2"
     """IBM Db2."""
-    POSTGRESQL = "postgresql"
+    POSTGRESQL: Final = "postgresql"
     """PostgreSQL."""
-    REDSHIFT = "redshift"
+    REDSHIFT: Final = "redshift"
     """Amazon Redshift."""
-    HIVE = "hive"
+    HIVE: Final = "hive"
     """Apache Hive."""
-    CLOUDSCAPE = "cloudscape"
+    CLOUDSCAPE: Final = "cloudscape"
     """Cloudscape."""
-    HSQLDB = "hsqldb"
+    HSQLDB: Final = "hsqldb"
     """HyperSQL DataBase."""
-    PROGRESS = "progress"
+    PROGRESS: Final = "progress"
     """Progress Database."""
-    MAXDB = "maxdb"
+    MAXDB: Final = "maxdb"
     """SAP MaxDB."""
-    HANADB = "hanadb"
+    HANADB: Final = "hanadb"
     """SAP HANA."""
-    INGRES = "ingres"
+    INGRES: Final = "ingres"
     """Ingres."""
-    FIRSTSQL = "firstsql"
+    FIRSTSQL: Final = "firstsql"
     """FirstSQL."""
-    EDB = "edb"
+    EDB: Final = "edb"
     """EnterpriseDB."""
-    CACHE = "cache"
+    CACHE: Final = "cache"
     """InterSystems Caché."""
-    ADABAS = "adabas"
+    ADABAS: Final = "adabas"
     """Adabas (Adaptable Database System)."""
-    FIREBIRD = "firebird"
+    FIREBIRD: Final = "firebird"
     """Firebird."""
-    DERBY = "derby"
+    DERBY: Final = "derby"
     """Apache Derby."""
-    FILEMAKER = "filemaker"
+    FILEMAKER: Final = "filemaker"
     """FileMaker."""
-    INFORMIX = "informix"
+    INFORMIX: Final = "informix"
     """Informix."""
-    INSTANTDB = "instantdb"
+    INSTANTDB: Final = "instantdb"
     """InstantDB."""
-    INTERBASE = "interbase"
+    INTERBASE: Final = "interbase"
     """InterBase."""
-    MARIADB = "mariadb"
+    MARIADB: Final = "mariadb"
     """MariaDB."""
-    NETEZZA = "netezza"
+    NETEZZA: Final = "netezza"
     """Netezza."""
-    PERVASIVE = "pervasive"
+    PERVASIVE: Final = "pervasive"
     """Pervasive PSQL."""
-    POINTBASE = "pointbase"
+    POINTBASE: Final = "pointbase"
     """PointBase."""
-    SQLITE = "sqlite"
+    SQLITE: Final = "sqlite"
     """SQLite."""
-    SYBASE = "sybase"
+    SYBASE: Final = "sybase"
     """Sybase."""
-    TERADATA = "teradata"
+    TERADATA: Final = "teradata"
     """Teradata."""
-    VERTICA = "vertica"
+    VERTICA: Final = "vertica"
     """Vertica."""
-    H2 = "h2"
+    H2: Final = "h2"
     """H2."""
-    COLDFUSION = "coldfusion"
+    COLDFUSION: Final = "coldfusion"
     """ColdFusion IMQ."""
-    CASSANDRA = "cassandra"
+    CASSANDRA: Final = "cassandra"
     """Apache Cassandra."""
-    HBASE = "hbase"
+    HBASE: Final = "hbase"
     """Apache HBase."""
-    MONGODB = "mongodb"
+    MONGODB: Final = "mongodb"
     """MongoDB."""
-    REDIS = "redis"
+    REDIS: Final = "redis"
     """Redis."""
-    COUCHBASE = "couchbase"
+    COUCHBASE: Final = "couchbase"
     """Couchbase."""
-    COUCHDB = "couchdb"
+    COUCHDB: Final = "couchdb"
     """CouchDB."""
-    COSMOSDB = "cosmosdb"
+    COSMOSDB: Final = "cosmosdb"
     """Microsoft Azure Cosmos DB."""
-    DYNAMODB = "dynamodb"
+    DYNAMODB: Final = "dynamodb"
     """Amazon DynamoDB."""
-    NEO4J = "neo4j"
+    NEO4J: Final = "neo4j"
     """Neo4j."""
-    GEODE = "geode"
+    GEODE: Final = "geode"
     """Apache Geode."""
-    ELASTICSEARCH = "elasticsearch"
+    ELASTICSEARCH: Final = "elasticsearch"
     """Elasticsearch."""
-    MEMCACHED = "memcached"
+    MEMCACHED: Final = "memcached"
     """Memcached."""
-    COCKROACHDB = "cockroachdb"
+    COCKROACHDB: Final = "cockroachdb"
     """CockroachDB."""
-    OPENSEARCH = "opensearch"
+    OPENSEARCH: Final = "opensearch"
     """OpenSearch."""
-    CLICKHOUSE = "clickhouse"
+    CLICKHOUSE: Final = "clickhouse"
     """ClickHouse."""
-    SPANNER = "spanner"
+    SPANNER: Final = "spanner"
     """Cloud Spanner."""
-    TRINO = "trino"
+    TRINO: Final = "trino"
     """Trino."""