+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/feature_flag_attributes.py
@@ -12,18 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Final
 
-FEATURE_FLAG_KEY = "feature_flag.key"
+FEATURE_FLAG_KEY: Final = "feature_flag.key"
 """
 The unique identifier of the feature flag.
 """
 
-FEATURE_FLAG_PROVIDER_NAME = "feature_flag.provider_name"
+FEATURE_FLAG_PROVIDER_NAME: Final = "feature_flag.provider_name"
 """
 The name of the service provider that performs the flag evaluation.
 """
 
-FEATURE_FLAG_VARIANT = "feature_flag.variant"
+FEATURE_FLAG_VARIANT: Final = "feature_flag.variant"
 """
 SHOULD be a semantic identifier for a value. If one is unavailable, a stringified version of the value can be used.
 Note: A semantic identifier, commonly referred to as a variant, provides a means