+++ shim/opentelemetry-opencensus-shim/src/opentelemetry/shim/opencensus/trace/_patch.py
@@ -19,9 +19,8 @@ from typing import Optional
 from opencensus.trace.span_context import SpanContext
 from opencensus.trace.tracer import Tracer
 from opencensus.trace.tracers.noop_tracer import NoopTracer
-
 from opentelemetry import trace
-from opentelemetry.shim.opencensus._shim_tracer import ShimTracer
+from opentelemetry.shim.opencensus.trace._shim_tracer import ShimTracer
 from opentelemetry.shim.opencensus.version import __version__
 
 _logger = getLogger(__name__)