+++ CHANGELOG.md
@@ -7,8 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Log a warning when a `LogRecord` in `sdk/log` has dropped attributes
+  due to reaching limits
+  ([#3946](https://github.com/open-telemetry/opentelemetry-python/pull/3946))
+- Fix RandomIdGenerator can generate invalid Span/Trace Ids
+  ([#3949](https://github.com/open-telemetry/opentelemetry-python/pull/3949))
+- Add Python 3.12 to tox
+  ([#3616](https://github.com/open-telemetry/opentelemetry-python/pull/3616))
+- Improve resource field structure for LogRecords
+  ([#3972](https://github.com/open-telemetry/opentelemetry-python/pull/3972))
+- Update Semantic Conventions code generation scripts:
+  - fix namespace exclusion that resulted in dropping  `os` and `net` namespaces.
+  - add `Final` decorator to constants to prevent collisions
+  - enable mypy and fix detected issues
+  - allow to drop specific attributes in preparation for Semantic Conventions v1.26.0
+  ([#3973](https://github.com/open-telemetry/opentelemetry-python/pull/3966))
+- Update semantic conventions to version 1.26.0.
+  ([#3964](https://github.com/open-telemetry/opentelemetry-python/pull/3964))
+- Use semconv exception attributes for record exceptions in spans
+  ([#3979](https://github.com/open-telemetry/opentelemetry-python/pull/3979))
+- Validate links at span creation
+  ([#3991](https://github.com/open-telemetry/opentelemetry-python/pull/3991))
+
 ## Version 1.25.0/0.46b0 (2024-05-30)
 
+- Fix _encode_events assumes events.attributes.dropped exists
+  ([#3965](https://github.com/open-telemetry/opentelemetry-python/pull/3965))
 - Fix class BoundedAttributes to have RLock rather than Lock
   ([#3859](https://github.com/open-telemetry/opentelemetry-python/pull/3859))
 - Remove thread lock by loading RuntimeContext explicitly.